Dedicated animation loader. Animations no longer belong to a model. They are passed when animated instead to prevent duplicate animation data. Animation controllers now do not own an animation. This is to allows multiple objects to use the same animation but with different play times and properties. Scene contains hold Effects, models, animations and animation controllers
